pyproject.toml
[project]
name = "marginalia"
version = "0.1.0"
description = "Margin notes for the videos you watch"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "google-genai>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
marginalia = "marginalia.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/marginalia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
